pyproject.toml
[project]
name = "bleled"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bleak>=2.1.1",
    "python-dotenv>=1.2.1",
]
